@@ -340,15 +67,20 @@
         "test": "node ./out/test/runTest.js",
         "test:rules": "node __test__/rule"
     },
+    "dependencies": {
+        "axios": "^0.19.0",
+		"pinyin": "^2.9.0"
+    },
     "devDependencies": {
         "@types/glob": "^7.1.1",
-        "@types/mocha": "^5.2.6",
-        "@types/node": "^10.12.21",
-        "@types/vscode": "^1.12.0",
-        "glob": "^7.1.4",
-        "mocha": "^6.1.4",
-        "typescript": "^3.3.1",
-        "tslint": "^5.12.1",
-        "vscode-test": "^1.0.2"
+		"@types/mocha": "^5.2.7",
+		"@types/node": "^12.11.7",
+        "@types/vscode": "^1.41.0",
+        "@types/pinyin": "^2.8.1",
+		"glob": "^7.1.5",
+		"mocha": "^6.2.2",
+		"tslint": "^5.20.0",
+		"typescript": "^3.6.4",
+		"vscode-test": "^1.2.2"
     }
 }
